moves domain layer to subpackage
