--- app/src/main/java/com/shevaalex/android/plugev/domain/Const.kt
@@ -1,3 +1,0 @@
-package com.shevaalex.android.plugev.domain
-
-const val API_RESULT_LIMIT = 250
