+++ app/src/main/java/com/shevaalex/android/plugev/domain/openchargemap/Const.kt
@@ -1,0 +1,3 @@
+package com.shevaalex.android.plugev.domain.openchargemap
+
+const val API_RESULT_LIMIT = 250
