+++ app/src/main/java/com/shevaalex/android/plugev/domain/openchargemap/usecase/GetChargeStationListUseCase.kt
@@ -1,8 +1,8 @@
-package com.shevaalex.android.plugev.domain.usecase
+package com.shevaalex.android.plugev.domain.openchargemap.usecase
 
-import com.shevaalex.android.plugev.domain.model.ChargingStation
-import com.shevaalex.android.plugev.domain.model.DataResult
-import com.shevaalex.android.plugev.domain.repository.ChargingStationRepository
+import com.shevaalex.android.plugev.domain.openchargemap.model.ChargingStation
+import com.shevaalex.android.plugev.domain.openchargemap.model.DataResult
+import com.shevaalex.android.plugev.domain.openchargemap.repository.ChargingStationRepository
 import dagger.hilt.android.scopes.ViewModelScoped
 import javax.inject.Inject
