+++ app/src/main/java/com/shevaalex/android/plugev/presentation/mapscreen/MarkerManager.kt
@@ -2,7 +2,7 @@ package com.shevaalex.android.plugev.presentation.mapscreen
 
 import com.google.android.gms.maps.model.LatLng
 import com.google.android.gms.maps.model.LatLngBounds
-import com.shevaalex.android.plugev.domain.model.ChargingStation
+import com.shevaalex.android.plugev.domain.openchargemap.model.ChargingStation
 import com.shevaalex.android.plugev.service.googlemap.PlugEvClusterManager
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ensureActive
