+++ app/src/test/java/com/shevaalex/android/plugev/data/network/openchargemap/model/ConnectionNetworkDtoTest.kt
@@ -2,7 +2,7 @@ package com.shevaalex.android.plugev.data.network.openchargemap.model
 
 import com.google.common.truth.Truth.assertThat
 import com.shevaalex.android.plugev.data.DataFactory
-import com.shevaalex.android.plugev.domain.model.Connection
+import com.shevaalex.android.plugev.domain.openchargemap.model.Connection
 import org.junit.Test
 
 class ConnectionNetworkDtoTest {
